# contracts/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Topos OSS plane",
    "description": "The self-hostable Topos plane — workspace-credential writes AND reads (one Bearer credential per enrolled device; membership is the authorization), plus the enrollment + governance surface. Every returned protocol outcome of an op_id-carrying write rides in a 200 body (the canonical JsonEnvelope + receipt); non-2xx is reserved for transport/auth/integrity faults.",
    "license": {
      "name": "Apache-2.0"
    },
    "version": "0.0.0"
  },
  "paths": {
    "/i/{token}": {
      "get": {
        "tags": [
          "enrollment"
        ],
        "operationId": "read_bootstrap",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "description": "The opaque `/i/<token>` one-time admin-claim token.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The claim bootstrap (workspace-to-be + plane API base; no bytes, no role, no trust root; enrollment_method \"admin_claim\", no skills). Content-negotiated: an Accept asking for JSON (or absent) gets this payload; anything else (curl `*/*`, a browser) gets a markdown agent-instruction document rendered from the same data, served as `text/plain` so browsers display it inline.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BootstrapData"
                }
              }
            }
          },
          "404": {
            "description": "No such claim, or it is revoked/consumed/expired (always the JSON envelope, any Accept).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (Retry-After header).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal store fault.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/workspaces/{ws}/policy/review-required": {
      "put": {
        "tags": [
          "policy"
        ],
        "operationId": "set_review_required",
        "parameters": [
          {
            "name": "ws",
            "in": "path",
            "description": "The workspace id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer <admin token> (the plane operator's secret).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyReviewRequiredRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "The policy was set (an idempotent write)."
          },
          "400": {
            "description": "Malformed body or workspace id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Admin token configured but missing/invalid on the request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No admin token is configured on this plane (the route is disabled).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal store fault.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Affected": {
        "type": "object",
        "description": "What an outcome refers to (every field optional — a policy flip has no skill/version).",
        "properties": {
          "proposal": {
            "type": [
              "string",
              "null"
            ]
          },
          "skill": {
            "type": [
              "string",
              "null"
            ]
          },
          "version": {
            "type": [
              "string",
              "null"
            ],
            "description": "A `version_id` (commit SHA-256, lowercase hex) when the outcome refers to a version."
          },
          "workspace": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "BootstrapData": {
        "type": "object",
        "description": "The payload an `/i/<token>` invite link resolves to — read once, before enrollment, to learn the\nworkspace + the plane API base to dial. No bytes, no role.",
        "required": [
          "schema_version",
          "invite",
          "plane",
          "workspace"
        ],
        "properties": {
          "invite": {
            "$ref": "#/components/schemas/BootstrapInvite",
            "description": "The invite's non-secret descriptor (consent posture, expiry)."
          },
          "offered_skills": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BootstrapSkill"
            },
            "description": "The skills the invite pre-offers (a name may be absent → the client shows the id). Empty for a\nmembership-only door."
          },
          "plane": {
            "$ref": "#/components/schemas/BootstrapPlane",
            "description": "The plane the workspace lives on — its API base and enrollment method."
          },
          "schema_version": {
            "type": "integer",
            "format": "int32",
            "description": "Always `1` for this contract version (the schema pins it `const`).",
            "minimum": 0
          },
          "workspace": {
            "$ref": "#/components/schemas/BootstrapWorkspace",
            "description": "The workspace the invite is for."
          }
        }
      },
      "BootstrapInvite": {
        "type": "object",
        "description": "The invite's non-secret descriptor — the consent posture a redeemer agrees to, never a role.",
        "required": [
          "token_id",
          "consent",
          "first_receive_auto_land"
        ],
        "properties": {
          "consent": {
            "$ref": "#/components/schemas/ConsentMode",
            "description": "The consent posture every first-receive follows."
          },
          "expires_at": {
            "type": [
              "string",
              "null"
            ],
            "description": "The invite's expiry as an RFC-3339 string, if it expires (`None` = never)."
          },
          "first_receive_auto_land": {
            "type": "boolean",
            "description": "Whether a first-received skill auto-lands. **ALWAYS false** — a received skill is offered, never\nsilently applied (a direct human yes is required)."
          },
          "token_id": {
            "type": "string",
            "description": "A stable, non-secret reference for this invite (the link token the client used to reach the\npayload). Empty for an admin-claim bootstrap — a claim token is a live one-time bearer capability\nthe server never echoes into a body; the claim client uses the token it parsed from the link."
          }
        }
      },
      "BootstrapPlane": {
        "type": "object",
        "description": "The plane a workspace lives on — its public base URL, deployment posture, and offered enrollment method.",
        "required": [
          "base_url",
          "deployment_mode",
          "enrollment_method"
        ],
        "properties": {
          "base_url": {
            "type": "string",
            "description": "The plane's public API base URL — the root the client dials for enrollment and sync after this\nbootstrap read. Always the plane itself, never a web front: a share link may ride another host,\nbut the client re-roots onto this value the moment the bootstrap is fetched."
          },
          "deployment_mode": {
            "$ref": "#/components/schemas/DeploymentMode",
            "description": "This plane's deployment posture."
          },
          "enrollment_method": {
            "type": "string",
            "description": "The enrollment method advertised to a bootstrapping device (e.g. `\"device_code\"` / `\"passcode\"`)."
          }
        }
      },
      "BootstrapSkill": {
        "type": "object",
        "description": "One skill an invite pre-offers — its id, plus an optional display name (names ride the invite; a missing\nname means the client shows the id).",
        "required": [
          "skill_id"
        ],
        "properties": {
          "name": {
            "type": [
              "string",
              "null"
            ],
            "description": "The skill's display name, if the invite carried one (else the client shows the id)."
          },
          "skill_id": {
            "type": "string",
            "description": "The offered skill id."
          }
        }
      },
      "BootstrapWorkspace": {
        "type": "object",
        "description": "The workspace an invite is for — its id, display name, and domain-verification state. No role (the role\nlives server-side on the pre-seeded member rows).",
        "required": [
          "workspace_id",
          "display_name",
          "verified_domain_status"
        ],
        "properties": {
          "display_name": {
            "type": "string",
            "description": "The workspace display name (shown on the verification page + in the agent's disclosure)."
          },
          "verified_domain": {
            "type": [
              "string",
              "null"
            ],
            "description": "The org-domain claim, if any."
          },
          "verified_domain_status": {
            "$ref": "#/components/schemas/VerifiedDomainStatus",
            "description": "The domain-verification state."
          },
          "workspace_id": {
            "type": "string",
            "description": "The workspace id."
          }
        }
      },
      "ConsentMode": {
        "type": "string",
        "description": "The consent posture of an invite — a CLOSED set. v0 is a single mode: a received skill is disclosed and\nawaits a direct human yes on first receive (never auto-landed).",
        "enum": [
          "direct_human_first_receive"
        ]
      },
...
      "DeploymentMode": {
        "type": "string",
        "description": "A plane's deployment posture — a CLOSED set. `cloud` requires a confirmed identity step at enrollment;\n`self_host` grants membership straight from a valid grant.",
        "enum": [
          "cloud",
          "self_host"
        ]
      },
...
      "JsonEnvelope": {
        "type": "object",
        "description": "The one common envelope every verb emits. Never prompts; prose (TTY) is rendered from the\nSAME typed value where practical.",
        "required": [
          "schema_version",
          "command",
          "ok"
        ],
        "properties": {
          "command": {
            "type": "string",
            "description": "The verb that produced this (`add`, `follow`, `pull`, `list`, `publish`, …)."
          },
          "data": {
            "description": "Command-specific payload (`{}` when empty)."
          },
          "error": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/WireError",
                "description": "The actionable failure detail (when `ok == false`)."
              }
            ]
          },
          "next_actions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NextAction"
            },
            "description": "Stable, machine-actionable next steps — each carries a complete argv."
          },
          "ok": {
            "type": "boolean"
          },
          "receipt": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Receipt",
                "description": "The durable idempotency record — present on **every** terminal op, including failures."
              }
            ]
          },
          "schema_version": {
            "type": "integer",
            "format": "int32",
            "description": "Always `1` for this contract version (the schema pins it `const`).",
            "minimum": 0
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "NextAction": {
        "type": "object",
        "description": "A machine-actionable next step. The `argv` is the ready-to-exec command; `code` lets an agent\nbranch on the known set and still pass through unknowns.",
        "required": [
          "code",
          "argv"
        ],
        "properties": {
          "argv": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "A complete argv array — execute as-is (no TTY parsing)."
          },
          "code": {
            "$ref": "#/components/schemas/ActionCode"
          }
        }
      },
...
      "PolicyReviewRequiredRequest": {
        "type": "object",
        "description": "`PUT /v1/workspaces/{ws}/policy/review-required` body — the self-host operator toggle for the\n`review-required` workspace policy (an idempotent set; JSON so the body stays extensible without a\npath-shape change). Authenticated by the plane's admin token, not a device credential; the route is\ninvisible (404) on a plane with no admin token configured.",
        "required": [
          "review_required"
        ],
        "properties": {
          "review_required": {
            "type": "boolean",
            "description": "The desired policy value: `true` gates any direct publish behind a reviewer's approval."
          }
        }
      },
...
      "VerifiedDomainStatus": {
        "type": "string",
        "description": "A workspace's domain-verification state — snake_case on the wire.",
        "enum": [
          "unverified",
          "pending",
          "verified"
        ]
      },
...
      "WireError": {
        "type": "object",
        "description": "The flat wire error (rich `thiserror` enums stay internal). Carries a\nstable code + retryability + the safe next actions; never raw SQL/git strings.",
        "required": [
          "code",
          "outcome",
          "retryable"
        ],
        "properties": {
          "affected": {
            "$ref": "#/components/schemas/Affected"
          },
          "code": {
            "type": "string",
            "description": "A stable, machine-branchable error code, distinct from (and finer than) `outcome`. This is an\n**open, additive** string vocabulary — it is intentionally NOT a closed enum: new codes are\nadded over time, and a consumer must treat an unrecognized `code` as the `outcome` it carries."
          },
          "context": {
            "description": "Structured context (no TTY prose to parse)."
          },
          "current_generation": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Generation"
              }
            ]
          },
          "expected_generation": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Generation"
              }
            ]
          },
          "next_actions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NextAction"
            }
          },
          "outcome": {
            "$ref": "#/components/schemas/TerminalOutcome"
          },
          "retryable": {
            "type": "boolean"
          }
        }
      },
...
    }
  },
...
}